Downgrades `scriptfilter`'s version for Alfred 3
Many of the workflow objects have had updates adding additional features
which make them no longer compatible with Alfred 3. Alfred 4 can happily
load Alfred 3 workflows, so if there is an absolute need for a workflow
to work in Alfred 3 and Alfred 4, one needs to create it with Alfred 3's
objects only.

Source:
https://www.alfredforum.com/topic/12994-solved-alfred-3-vs-4-compatibility/?do=findComment&comment=67907